* src/qemu_conf.c src/qemu_conf.h src/qemu_driver.c: applied patch   from Cole Robinson implementing memory operations for qemu driver Daniel

+++ src/qemu_conf.h
@@ -193,8 +193,8 @@ struct qemud_vm_def {
     unsigned char uuid[VIR_UUID_BUFLEN];
     char name[QEMUD_MAX_NAME_LEN];
 
-    int memory;
-    int maxmem;
+    unsigned long memory;
+    unsigned long maxmem;
     int vcpus;
 
     int noReboot;
